Mise à Jour 04/04/2022
Correction mineure + Ajout et modification Projet Fonctionnel + Budget delete fonctionnel

+++ src/Controller/BudgetController.php
@@ -6,6 +6,7 @@ use App\Entity\Budget;
 use App\Entity\Project;
 use App\Form\BudgetType;
 use App\Repository\BudgetRepository;
+use App\Repository\ProjectRepository;
 use Doctrine\ORM\EntityManagerInterface;
 use Sensio\Bundle\FrameworkExtraBundle\Configuration\Entity;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
@@ -68,23 +69,53 @@ class BudgetController extends AbstractController
         ]);
     }
 
+    #[Route('/{id}/delete_view/{projectId}', name: 'app_budget_delete_view', methods: ['GET'])]
+    /**
+     * @Entity("project", expr="repository.find(projectId)")
+     */
+    public function delete_view(Budget $budget, Project $project): Response
+    {
+        return $this->render('budget/delete_view.html.twig', [
+            'budget' => $budget,
+            'project' => $project,
+        ]);
+    }
+
     #[Route('/{id}/{projectId}', name: 'app_budget_delete', methods: ['POST'])]
     /**
      * @Entity("project", expr="repository.find(projectId)")
      */
     public function delete(Request $request, Budget $budget,
                            EntityManagerInterface $entityManager,
-                           Project $project): Response
+                           Project $project, ProjectRepository $projectRepo): Response
     {
         if ($this->isCsrfTokenValid('delete'.$budget->getId(), $request->request->get('_token'))) {
-            $entityManager->remove($budget);
-            $entityManager->flush();
-            $this->addFlash(
-                'SuccessDeleteBudget',
-                'Votre budget a été supprimé !'
-            );
+
+            $projectsRelated = $projectRepo->findByBudget($budget->getId());
+
+            if ($projectsRelated == null) {
+                $entityManager->remove($budget);
+                $entityManager->flush();
+                $this->addFlash(
+                    'SuccessDeleteBudget',
+                    'Votre budget a été supprimé !'
+                );
+            }
+            else {
+                $this->addFlash(
+                    'deleteDangerBudget',
+                    'Erreur, impossible de supprimer ce budget.
+                Veuillez vérifier que tous les projets ayant assignés ce budget soient changés.'
+                );
+                return $this->redirectToRoute('app_budget_delete_view',
+                    ['id' => $budget->getId(),
+                    'projectId' => $project->getId()])
+                ;
+
+            }
+            return $this->redirectToRoute('app_project_show', ['projectId' => $project->getId()], Response::HTTP_SEE_OTHER);
+
         }
 
-        return $this->redirectToRoute('app_project_show', ['projectId' => $project->getId()], Response::HTTP_SEE_OTHER);
     }
 }
